+++ CHANGES.md
@@ -22,7 +22,8 @@
 ## v0.1.15
 
  - [x] Kinetic rates are now computed from model instead of deriving solutions 
- - [x] Refactored namespace to be more sensical 
+ - [x] Refactored namespace to be more sensical
+ - [x] Added Chromatogram Solver to detect peaks 
 
 ## v0.1.14